Added support for key value enums.
Before this patch, only a list of string could be provided to the enum
attribute, leading to a list of options with the same name and value for
each entry of the list.

With this patch, you can still provide a list or if you want to gain
control over the name and value of the options, you can provide a
dictionnary whose key will become the name and value, well, the
displayed value.

eg:
	{
	  "type": "object",
	  "properties": {
	    "number":      { "type": "number" },
	    "street_name": { "type": "string" },
	    "street_type": {
	      "type": "string",
	      "enum": ["Street", "Avenue", "Boulevard"]
	    }
	  }
	}
    OR
        {
          "type": "object",
          "properties": {
	    "number":      { "type": "number" },
	    "street_name": { "type": "string" },
	    "street_type": {
	      "type": "string",
	      "enum": {"street":"A Street", "avenue":"A nice avenue"}
	    }
          }
        }